python/requirements.txt
pyautogui~=0.9.54
opencv-python<=4.6.0.66
opencv-contrib-python<=4.6.0.66
numpy~=1.26.0
#Pillow~=9.5.0 图像处理库,用于图片读取,处理和保存
#pywin32
paddlepaddle
paddleocr~=2.6.1.3
gradio~=3.45.0
loguru~=0.7.2
pillow~=9.5.0
matplotlib~=3.8.0
memory_profiler
mss~=9.0.1
pyyaml~=6.0.1